+++ tests/test_skoolasm.py
@@ -2830,22 +2830,6 @@ class AsmWriterTest(SkoolKitTestCase, CommonSkoolMacroTest):
         """
         self._test_asm(skool, exp_asm)
 
-    def test_push_snapshot_keeps_original_in_place(self):
-        writer = self._get_writer(snapshot=[0])
-        snapshot = writer.snapshot
-        writer.push_snapshot()
-        writer.snapshot[0] = 1
-        self.assertEqual(snapshot[0], 1)
-
-    def test_pop_snapshot_modifies_snapshot_in_place(self):
-        writer = self._get_writer(snapshot=[0])
-        snapshot = writer.snapshot
-        writer.snapshot[0] = 1
-        writer.push_snapshot()
-        writer.snapshot[0] = 2
-        writer.pop_snapshot()
-        self.assertEqual(snapshot[0], 1)
-
     def test_last_instruction_line_with_no_address(self):
         skool = """
             @start