+++ tests/test_skoolhtml.py
@@ -1396,22 +1396,6 @@ class MethodTest(HtmlWriterTestCase):
         """
         self._test_format_template(ref, 'test', fields, exp_output)
 
-    def test_push_snapshot_keeps_original_in_place(self):
-        writer = self._get_writer(snapshot=[0])
-        snapshot = writer.snapshot
-        writer.push_snapshot()
-        writer.snapshot[0] = 1
-        self.assertEqual(snapshot[0], 1)
-
-    def test_pop_snapshot_modifies_snapshot_in_place(self):
-        writer = self._get_writer(snapshot=[0])
-        snapshot = writer.snapshot
-        writer.snapshot[0] = 1
-        writer.push_snapshot()
-        writer.snapshot[0] = 2
-        writer.pop_snapshot()
-        self.assertEqual(snapshot[0], 1)
-
     def test_get_snapshot_name(self):
         writer = self._get_writer(snapshot=[])
         names = ['snapshot1', 'next', 'final']