+++ sphinx/source/changelog.rst
@@ -24,6 +24,8 @@ Changelog
   of the tape's tones, pulse sequences and data blocks)
 * Added the ``--tape-skip`` option to :ref:`tapinfo.py` (for skipping one or
   more blocks on a tape)
+* Fixed how :ref:`PUSHS` and :ref:`POPS` copy and restore the RAM bank mapped
+  to 0xC000-0xFFFF in a 128K memory snapshot
 * Fixed how the ``--basic`` option of :ref:`tapinfo.py` handles a block that
   has no data
 * Fixed how the ``--tape-stop`` option of :ref:`tap2sna.py` works with two